+++ tiled/access_control/access_policies.py
@@ -1,12 +1,9 @@
 import logging
 import os
-import sqlite3
-from contextlib import closing
-from functools import partial
 
-from .queries import AccessBlobFilter, In, KeysFilter
+from ..queries import AccessBlobFilter
+from ..utils import Sentinel, import_object
 from .scopes import ALL_SCOPES, PUBLIC_SCOPES
-from .utils import Sentinel, SpecialUsers, import_object
 
 ALL_ACCESS = Sentinel("ALL_ACCESS")
 NO_ACCESS = Sentinel("NO_ACCESS")
@@ -31,153 +28,6 @@ class DummyAccessPolicy:
 
     async def filters(self, node, principal, authn_scopes, scopes):
         return []
-
-
-class SimpleAccessPolicy:
-    """
-    A mapping of user names to lists of entries they have access to.
-
-    This simple policy does not provide fine-grained control of scopes.
-    Any restriction on scopes is applied the same to all users, except
-    for an optional list of 'admins'.
-
-    This is used in the test suite; it may be suitable for very simple
-    deployments.
-
-    >>> SimpleAccessPolicy({"alice": ["A", "B"], "bob": ["B"]}, provider="toy")
-    """
-
-    ALL = ALL_ACCESS
-
-    def __init__(
-        self, access_lists, *, provider, key=None, scopes=None, public=None, admins=None
-    ):
-        self.access_lists = {}
-        self.provider = provider
-        self.key = key
-        self.scopes = scopes if (scopes is not None) else ALL_SCOPES
-        self.public = set(public or [])
-        self.admins = set(admins or [])
-        for key, value in access_lists.items():
-            if isinstance(value, str):
-                value = import_object(value)
-            self.access_lists[key] = value
-
-    def _get_id(self, principal):
-        # Get the id (i.e. username) of this Principal for the
-        # associated authentication provider.
-        for identity in principal.identities:
-            if identity.provider == self.provider:
-                id = identity.id
-                break
-        else:
-            raise ValueError(
-                f"Principcal {principal} has no identity from provider {self.provider}. "
-                f"Its identities are: {principal.identities}"
-            )
-        return id
-
-    async def allowed_scopes(self, node, principal, authn_scopes):
-        # If this is being called, filter_access has let us get this far.
-        if principal is SpecialUsers.public:
-            allowed = PUBLIC_SCOPES
-        elif principal.type == "service":
-            allowed = self.scopes
-        elif self._get_id(principal) in self.admins:
-            allowed = ALL_SCOPES
-        # The simple policy does not provide for different Principals to
-        # have different scopes on different Nodes. If the Principal has access,
-        # they have the same hard-coded access everywhere.
-        else:
-            allowed = self.scopes
-        return allowed
-
-    async def filters(self, node, principal, authn_scopes, scopes):
-        queries = []
-        query_filter = KeysFilter if not self.key else partial(In, self.key)
-        if principal is SpecialUsers.public:
-            queries.append(query_filter(self.public))
-        else:
-            # Services have no identities; just use the uuid.
-            if principal.type == "service":
-                id = str(principal.uuid)
-            else:
-                id = self._get_id(principal)
-            if id in self.admins:
-                return queries
-            if not scopes.issubset(self.scopes):
-                return NO_ACCESS
-            access_list = self.access_lists.get(id, [])
-            if not ((principal is SpecialUsers.admin) or (access_list == self.ALL)):
-                try:
-                    allowed = set(access_list or [])
-                except TypeError:
-                    # Provide rich debugging info because we have encountered a confusing
-                    # bug here in a previous implementation.
-                    raise TypeError(
-                        f"Unexpected access_list {access_list} of type {type(access_list)}. "
-                        f"Expected iterable or {self.ALL}, instance of {type(self.ALL)}."
-                    )
-                queries.append(query_filter(allowed))
-        return queries
-
-
-class AccessTagsParser:
-    @classmethod
-    def from_uri(cls, uri):
-        db = sqlite3.connect(f"{uri}?ro", uri=True, check_same_thread=False)
-        return cls(db)
-
-    def __init__(self, db):
-        self.db = db
-
-    def is_tag_defined(self, name):
-        with closing(self.db.cursor()) as cursor:
-            cursor.execute("SELECT 1 FROM tags WHERE name = ?;", (name,))
-            row = cursor.fetchone()
-            found_tagname = bool(row)
-        return found_tagname
-
-    def get_public_tags(self):
-        with closing(self.db.cursor()) as cursor:
-            cursor.execute("SELECT name FROM public_tags;")
-            public_tags = {name for (name,) in cursor.fetchall()}
-        return public_tags
-
-    def get_scopes_from_tag(self, tagname, username):
-        with closing(self.db.cursor()) as cursor:
-            cursor.execute(
-                "SELECT scope_name FROM user_tag_scopes WHERE tag_name = ? AND user_name = ?;",
-                (tagname, username),
-            )
-            user_tag_scopes = {scope for (scope,) in cursor.fetchall()}
-        return user_tag_scopes
-
-    def is_tag_owner(self, tagname, username):
-        with closing(self.db.cursor()) as cursor:
-            cursor.execute(
-                "SELECT 1 FROM user_tag_owners WHERE tag_name = ? AND user_name = ?;",
-                (tagname, username),
-            )
-            row = cursor.fetchone()
-            found_owner = bool(row)
-        return found_owner
-
-    def is_tag_public(self, name):
-        with closing(self.db.cursor()) as cursor:
-            cursor.execute("SELECT 1 FROM public_tags WHERE name = ?;", (name,))
-            row = cursor.fetchone()
-            found_public = bool(row)
-        return found_public
-
-    def get_tags_from_scope(self, scope, username):
-        with closing(self.db.cursor()) as cursor:
-            cursor.execute(
-                "SELECT tag_name FROM user_tag_scopes WHERE user_name = ? AND scope_name = ?;",
-                (username, scope),
-            )
-            user_scope_tags = {tag for (tag,) in cursor.fetchall()}
-        return user_scope_tags
 
 
 class TagBasedAccessPolicy:
@@ -398,7 +248,9 @@ class TagBasedAccessPolicy:
         elif self._is_admin(authn_scopes):
             allowed = self.scopes
         else:
-            if principal.type == "service":
+            if principal is None:
+                identifier = None
+            elif principal.type == "service":
                 identifier = str(principal.uuid)
             else:
                 identifier = self._get_id(principal)
@@ -415,7 +267,8 @@ class TagBasedAccessPolicy:
                             continue
                     elif not self.is_tag_defined(tag):
                         continue
-                    tag_scopes = self.get_scopes_from_tag(tag, identifier)
+                    if identifier is not None:
+                        tag_scopes = self.get_scopes_from_tag(tag, identifier)
                     allowed.update(
                         tag_scopes if tag_scopes.issubset(self.scopes) else set()
                     )
@@ -431,16 +284,22 @@ class TagBasedAccessPolicy:
         if not scopes.issubset(self.scopes):
             return NO_ACCESS
 
-        if principal.type == "service":
-            identifier = str(principal.uuid)
-        elif self._is_admin(authn_scopes):
-            return queries
+        tag_list = set()
+        if principal is None:
+            identifier = None
         else:
-            identifier = self._get_id(principal)
+            if principal.type == "service":
+                identifier = str(principal.uuid)
+            elif self._is_admin(authn_scopes):
+                return queries
+            else:
+                identifier = self._get_id(principal)
+            tag_list.update(
+                set.intersection(
+                    *[self.get_tags_from_scope(scope, identifier) for scope in scopes]
+                )
+            )
 
-        tag_list = set.intersection(
-            *[self.get_tags_from_scope(scope, identifier) for scope in scopes]
-        )
         tag_list.update(
             set.intersection(
                 *[