+++ tiled/server/app.py
@@ -48,7 +48,7 @@ from ..media_type_registration import (
     default_deserialization_registry,
     default_serialization_registry,
 )
-from ..utils import SHARE_TILED_PATH, Conflicts, SpecialUsers, UnsupportedQueryType
+from ..utils import SHARE_TILED_PATH, Conflicts, UnsupportedQueryType
 from ..validation_registration import ValidationRegistry, default_validation_registry
 from .compression import CompressionMiddleware
 from .router import get_router
@@ -360,8 +360,6 @@ def build_app(
     ) -> JSONResponse:
         # The current_principal_logging_filter middleware will not have
         # had a chance to finish running, so set the principal here.
-        principal = getattr(request.state, "principal", None)
-        current_principal.set(principal)
         return await http_exception_handler(
             request,
             HTTPException(
@@ -555,7 +553,10 @@ def build_app(
             # registry, keyed on database_settings, where can be retrieved by
             # the Dependency get_database_session.
             engine = open_database_connection_pool(settings.database_settings)
-            if not engine.url.database:
+            if (
+                not engine.url.database
+                or engine.url.query.get("mode", None) == "memory"
+            ):
                 # Special-case for in-memory SQLite: Because it is transient we can
                 # skip over anything related to migrations.
                 await initialize_database(engine)
@@ -841,7 +842,7 @@ def build_app(
 
     @app.middleware("http")
     async def current_principal_logging_filter(request: Request, call_next):
-        request.state.principal = SpecialUsers.public
+        request.state.principal = None
         response = await call_next(request)
         current_principal.set(request.state.principal)
         return response